feat: Date Formatter 메서드 구현
- format(data:) 메서드 구현
    - Date Formatter로 날짜 형식 지정
- DiarySample 데이터 매핑 모델의 타입 수정
    - Int -> Double로 변경

+++ Diary/DiarySample.swift
@@ -12,7 +12,7 @@ struct DiarySample: Decodable, Hashable {
     
     let title: String
     let body: String
-    let createdAt: Int
+    let createdAt: Double
     
     enum CodingKeys: String, CodingKey {
         case title, body
